meme-maker-backend/requirements.txt
Flask==2.2.5
Pillow==9.0.0
torch==2.0.0           
accelerate>=0.26.0
python-dotenv==0.21.0
openai==0.28
numpy<2.0.0 
flask-cors           
attrs

